feat,refactor: save settings to shared_preferences instance, also move note_provider to providers/ and auth to services/
