+++ frontend/lib/providers/note_provider.dart
