+++ frontend/windows/flutter/generated_plugins.cmake
@@ -7,6 +7,7 @@ list(APPEND FLUTTER_PLUGIN_LIST
   file_selector_windows
   firebase_auth
   firebase_core
+  flutter_tts
 )
 
 list(APPEND FLUTTER_FFI_PLUGIN_LIST
